--- a/examples/secrets.rs
+++ b/examples/secrets.rs
@@ -1,0 +1,7 @@
+extern crate dyon;
+
+use dyon::{error, run};
+
+fn main() {
+    error(run("source/secrets.dyon"));
+}