--- a/source/test.dyon
+++ b/source/test.dyon
@@ -1,35 +1,13 @@
 
 fn main() {
     /*
-    list := [1, 2, 3]
-    x := any i { list[i] > 1 }
-    if x {
-        println(why(x))
-    }
-
-    list := [[1, 2], [3, 4]]
-    x := any i, j { list[i][j] == 2 }
-    if x {
-        println(why(x))
-    }
-
-    list := [1]
-    x := any i { list[i] > 0 }
-    println(why(x))
+    list := [2, 1, 0]
+    x := min i { list[i] }
+    println(where(x))
     */
 
-    list := [[4, 4], [3, 4]]
-    x := any i, j { list[i][j] != 4 }
-    println(why(x))
-
-    /*
-    list := [2, 1, 3]
-    x := !all i { list[i] > 1 }
-    println(why(x))
-    */
-
-    /*
-    list := [[3, 4], [1, 2]]
-    println(why(!all i, j { list[i][j] > 2 }))
-    */
+    list := [[1, 2], [3, 2]]
+    x := max i { min j { list[i][j] } }
+    println(x)
+    println(where(x))
 }