--- a/source/test.dyon
+++ b/source/test.dyon
@@ -1,27 +1,9 @@
 
 fn main() {
-    /*
-    list := [[3, 3], [2, 3]]
-    x := !all i { min j {list[i][j]} > 2}
-    println(x)
-    println(why(x))
-    */
+    x := any i 3 { i == 1 }
+    println(foo(x))
+}
 
-    /*
-    list := [[1, 2], [3, 4]]
-    x := any i { all j {list[i][j] < 4} == false }
-    println(why(x))
-    */
-
-    /*
-    list := [1, 2, 3]
-    x := any i { list[i] == 2 } || all i { list[i] < 3 }
-    println(x)
-    println(why(x))
-    */
-
-    list := [1, 3, 2]
-    x := max i { list[i] } / sum i { list[i] }
-    println(x)
-    println(where(x))
+fn foo(x: 'return) -> [f64] {
+    return why(x)
 }