--- a/source/test.dyon
+++ b/source/test.dyon
@@ -2,15 +2,21 @@
 fn main() {
     list := [1, 2, 3]
     reason := "in the list!"
-    x := any i {
+    x := false
+    x = any i {
         x := explain_why(list[i] == 1, reason)
         explain_why(x, reason)
     }
     println(why(x))
 
-    x := max i {
+    /*
+    list := [1, 2, 3]
+    reason := "in the list!"
+    x := 0
+    x = max i {
         x := explain_where(list[i], reason)
         explain_where(x, reason)
     }
     println(where(x))
+    */
 }