--- a/source/test.dyon
+++ b/source/test.dyon
@@ -1,5 +1,6 @@
 
 fn main() {
+    /*
     list := [1, 2, 3]
     reason := "in the list!"
     x := false
@@ -8,6 +9,7 @@ fn main() {
         explain_why(x, reason)
     }
     println(why(x))
+    */
 
     /*
     list := [1, 2, 3]
@@ -19,4 +21,55 @@ fn main() {
     }
     println(where(x))
     */
+
+    /*
+    new_emails := ["Message from John"]
+    x := got_new_email(new_emails)
+    if x {
+        println(why(x))
+    }
+    */
+
+    /*
+    list := [1, 2, 3]
+    x := foo(list)
+    if x {
+        println(why(x))
+    }
+    */
+
+    /*
+    list := [1, 2, 3]
+    x := foo(list)
+    println(where(x))
+    */
+
+    /*
+    reason := "just because"
+    x := explain_why((1+1)==2, reason)
+    println(why(x)) // prints `just because`
+    */
+
+    reason := "the number is 2"
+    x := explain_where(2, reason)
+    println(where(x)) // prints `the number is 2`
 }
+
+/*
+fn got_new_email(new_emails: 'return [str]) -> bool {
+    if len(new_emails) == 0 { return false }
+    else { return explain_why(true, new_emails[0]) }
+}
+*/
+
+/*
+fn foo(list) -> {
+    return any i { list[i] == 2 }
+}
+*/
+
+/*
+fn foo(list) -> {
+    return max i { list[i] }
+}
+*/