--- a/src/intrinsics/meta.rs
+++ b/src/intrinsics/meta.rs
@@ -28,8 +28,8 @@ fn load_metarules_data(meta: &str, s: &str, file: &str, d: &str) -> Result<Vec<V
     let end: Arc<String> = Arc::new("end".into());
     for range_token in &tokens {
         let mut data = vec![];
-        data.push(Variable::F64(range_token.offset as f64));
-        data.push(Variable::F64(range_token.length as f64));
+        data.push(Variable::f64(range_token.offset as f64));
+        data.push(Variable::f64(range_token.length as f64));
         match &range_token.data {
             &MetaData::Bool(ref name, val) => {
                 data.push(Variable::Text(b.clone()));
@@ -44,7 +44,7 @@ fn load_metarules_data(meta: &str, s: &str, file: &str, d: &str) -> Result<Vec<V
             &MetaData::F64(ref name, val) => {
                 data.push(Variable::Text(n.clone()));
                 data.push(Variable::Text(name.clone()));
-                data.push(Variable::F64(val));
+                data.push(Variable::f64(val));
             }
             &MetaData::StartNode(ref name) => {
                 data.push(Variable::Text(start.clone()));
@@ -215,7 +215,7 @@ pub fn json_from_meta_data(data: &Vec<Variable>) -> Result<String, io::Error> {
                         }
                     }
                     "f64" => {
-                        if let &Variable::F64(val) = &arr[4] {
+                        if let &Variable::F64(val, _) = &arr[4] {
                             try!(write_string(&mut w, name));
                             try!(write!(w, ":{}", val));
                         }