--- a/src/intrinsics/mod.rs
+++ b/src/intrinsics/mod.rs
@@ -36,6 +36,11 @@ pub fn standard(f: &mut HashMap<Arc<String>, PreludeFunction>) {
         tys: vec![Type::F64],
         ret: Type::array()
     });
+    f.insert(Arc::new("explain_why".into()), PreludeFunction {
+        lts: vec![Lt::Default, Lt::Return],
+        tys: vec![Type::Bool, Type::Any],
+        ret: Type::Bool
+    });
     sarg(f, "println", Type::Any, Type::Void);
     sarg(f, "print", Type::Any, Type::Void);
     sarg(f, "clone", Type::Any, Type::Any);
@@ -441,6 +446,19 @@ pub fn call_standard(
                     &rt.expected(x, "f64"), rt))
             }));
             rt.stack.push(v);
+            rt.pop_fn(call.name.clone());
+            Expect::Something
+        }
+        "explain_why" => {
+            rt.push_fn(call.name.clone(), 0, None, st + 1, lc, cu);
+            let why = rt.stack.pop().expect(TINVOTS);
+            let val = rt.stack.pop().expect(TINVOTS);
+            let val = match rt.resolve(&val) {
+                &Variable::Bool(val, _) => val,
+                x => return Err(module.error(call.args[0].source_range(),
+                    &rt.expected(x, "bool"), rt))
+            };
+            rt.stack.push(Variable::Bool(val, Some(Box::new(vec![why]))));
             rt.pop_fn(call.name.clone());
             Expect::Something
         }